Revert "clean up service methods with .orElseThrow()"
This reverts commit 612af414fd7cb23fe0658ce06aa2f82d015b37a0.

+++ src/main/java/com/colinven/blog/service/BlogService.java
@@ -32,8 +32,10 @@ public class BlogService {
     }
 
     public BlogResponse getBlog(Long id) {
-        Blog blog = blogRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Blog with id: " + id + " not found."));
+        Blog blog = blogRepository.findById(id).orElse(null);
+        if (blog == null) {
+            throw new ResourceNotFoundException("Blog with id " + id + " not found");
+        }
         return new BlogResponse(blog);
     }
 
@@ -48,9 +50,10 @@ public class BlogService {
     }
 
     public BlogResponse editBlog(Long id, BlogRecord blogRecord) {
-        Blog blog = blogRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Blog with id: " + id + " not found."));
-
+        Blog blog = blogRepository.findById(id).orElse(null);
+        if (blog == null) {
+            throw new ResourceNotFoundException("Blog with id " + id + " not found");
+        }
         blog.setTitle(blogRecord.title());
         blog.setContent(blogRecord.content());
         blog.setCategory(blogRecord.category());
@@ -61,8 +64,10 @@ public class BlogService {
     }
 
     public void deleteBlog(Long id) {
-        Blog blog = blogRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Blog with id: " + id + " not found."));
+        Blog blog = blogRepository.findById(id).orElse(null);
+        if (blog == null) {
+            throw new ResourceNotFoundException("Blog with id " + id + " not found");
+        }
         blogRepository.delete(blog);
     }
